0322 - board JPA 서비스 로직 수정중

+++ javajava/src/main/java/com/example/javajava/board/dto/FileDto.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class FileDto {
-    private int idx;
-    private int boardIdx;
+    private Long idx;
+    private Long boardIdx;
     private String originalFileName;
     private String storedFilePath;
-    private long fileSize;
+    private Long fileSize;
 }
